+++ asset_allocation/strategies/relative_value.py
@@ -129,7 +129,7 @@ class RelativeValue(strategy.Strategy):
         ss = np.arange(self.num_total_assets)
         return ss[
             ~np.isnan(self.latest_return)
-            & self.index_comp
+            & (self.index_comp > 0)
             & (self.market_caps > self.market_cap_threshold)
             & (self.divyield > self.div_yield_threshold)
             & (self.roe > self.roe_threshold)